Support block parsing for classic blocks
Classic blocks are not wrapped with the usual `<!-- wp:block -->`
comments. But it's parsed by WP (in `parse_blocks`) which returns a block
without name.

In order to also process those classic blocks in the translation editor,
we will temporarily give it a name so it can be handled as any other
block.

At rendering time, we will process it as a raw HTML and apply `wpautop`
as it depends if it was translated with the visual tab or text tab.

Unfortunately, I could not add test because the `WP_Block_Parser_Block`
class is hardcoded inside our code.

https://onthegosystems.myjetbrains.com/youtrack/issue/wpmlcore-6075