Force to update the post content in the target language
We have a filter WPML_Term_Adjust_Id::filter which is messing the terms
when we are saving the post content. Sometimes it converts the
translated terms into the source ones (because the current lang is the
original one).

The filter is relying on several debug backtrace conditions which are
impacted depending on the activated page builders (e.g. we have a
problem when Enfold + Gutenberg are enabled, but not when only Gutenberg
is enabled).

By saving in the target lang, we ensure the terms are not messed even if
it's filtered.

Moreover, at this step we should not have nested post save so we should
not have side effects.

https://onthegosystems.myjetbrains.com/youtrack/issue/wpmlcore-6221