Include the reusable string packages in the translation job of the document it's inserted in.
When the reusable block is saved, strings will be created in a package
attached to the reusable block ID. We will add this package to the main
package of strings attached to the post we are translating.

In order to have a better organization of the code and decoupling, I
created the `Integration_Composite` class, so we can load several
integration classes.

https://onthegosystems.myjetbrains.com/youtrack/issue/wpmlcore-6563