Convert the reusable block ID on the fly when it's rendered
I decided to remove the code to convert the ID inside the translated
post because we would keep an invalid block ID if one day this
translated block is removed independently. Also, we'll have a consistent
way to convert the block accross all posts.

In order to have a better organization of the code and decoupling, I
created the `Integration_Composite` class, so we can load several
integration classes.

https://onthegosystems.myjetbrains.com/youtrack/issue/wpmlcore-6565