Translate inner content strings in a block with inner blocks
Until now, we where handling blocks with inner blocks as pure layouts
(like columns for instance). The content around the inner blocks was not
supposed to be translate as it was only html tags for the layout.

But some blocks are also having translatable strings around the inner
blocks.

I could not use the DOM to parse the inner content because it's made of
HTML extracts which are invalid taken independently. Using the DOM
parser will also fix those extracts to a valid HTML which is not what we
want here.

https://onthegosystems.myjetbrains.com/youtrack/issue/wpmlcore-6627